// src/types/pokemonAPI.ts
export interface Result {
    name: string;
    url:  string;
}